feat: Geocoding article locations
Created file geocode-location.ipynb to work on harvesting the human coded addresses in the news article and prepare them to be used in one of the available geocding services. Goal is to create a map of where the articles are identifying the locations of the crimes being reported.

Open Issue: #23