Switch to uv for project management
This change is motivated by the need to update this project's required
Python version (3.8 reached end-of-life on 2024-10-07).

uv was chosen because it provides good dependency resolution and easy
management of virtual environments.

Some trial and error was required to determine the proper versions for
project dependencies (since none were specified prior to this commit).
TensorFlow was pinned at 2.15.1 because versions 2.16 and up require
Keras version 3, which doesn't work with this project. Other versions
were resolved automatically by uv.

A guide to using uv in GitHub actions can be found at
<https://docs.astral.sh/uv/guides/integration/github/>.